pyproject.toml
[project]
name = "xtalbake"
version = "0.1.0"
license = {file = "LICENSE"}
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "peter.barrow.93@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "matplotlib",
    "pyqt6",
    "pyserial",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
xtalbake-ui = "xtalbake.ui:main"
xtalbake-gui = "xtalbake.xtalbake_gui:main"
xtalbake-discover = "xtalbake.device_discovery:main"
